# Change implementation of acl_hal_mmd_get_timestamp() to use std::chro…