@huangbri 2022-03-18
Current benchmarks:

old: 3278.61 ms
new: 3457.42 ms

old (no profiling): 3359.73 ms
new (no profiling): 3171.21 ms

Looks like there is random variation. Going to do multiple runs and average out.